put the enums in their own enum package and made a test for getting the nail enum objects from the arraylist in TestArrayList

+++ swe400_01_SingleTable/src/swe400_01_SingleTable/DatabaseGateway.java
@@ -60,7 +60,7 @@ public class DatabaseGateway
 			 rs1.setPrice(rs.getInt("price"));
 			 rs1.setDescription(rs.getString("description"));
 			 rs1.setBatteryPowered(rs.getBoolean("batteryPowered"));
-			 rs1.setLength(rs.getLong("length"));
+			 rs1.setLength(rs.getDouble("length"));
 			 rs1.setNumberInStrip(rs.getInt("numberInStrip"));
 			 rs1.setNumberInBox(rs.getInt("numberInBox"));
 			 rs1.setClassName(rs.getString("className"));
